Add files via upload
Lifes work... escape all multiplexes xor n ur /*/*/*/*/ xD
basically and honnestly some unknown force got the equation for this damn pattern faded in the myriad of equation backup, i may have lost it, but this was/it/will be my portal to the next layered multitransversal trolling riddle... god is a troll! but that entity makes this struggle worth considering to endure. that's it i gotta stop speaking my mind raw, like this, need to get laid actually... yeah good this repo is private, or something. haha xD take care <3